+++ webview/src/components/TsResultModal/ResultRenameDrawer/ResultRenameDrawer.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import styles from './ResultRenameDrawer.less';
 import { Typography, Collapse, Drawer, DrawerProps, Space, Button, Descriptions, Empty, Form, theme, Divider, Tooltip, Modal } from 'antd';
-import { ApiGroupNameMapping, NameMappingByGroup, RenameMapping } from '../../../../../src/types';
+import { ApiGroupNameMapping, NameMappingByGroup, RenameMapping, ServiceMapInfoYAMLJSONType } from '../../../../../src/types';
 import { CheckCircleFilled, DownOutlined, ExclamationCircleOutlined, SyncOutlined } from '@ant-design/icons';
 import MethodTag from '../../MethodTag';
 import { isPlainObject } from 'lodash-es';
@@ -61,10 +61,20 @@ export interface ResultRenameDrawerProps extends Omit<DrawerProps, 'onClose'> {
   allDefNameMapping?: Record<string, string>;
   onClose?: () => void;
   onAfterRenameTs?: (result: V2TSGenerateResult) => void;
+  localServiceInfo?: ServiceMapInfoYAMLJSONType[];
 }
 
 const ResultRenameDrawer: React.FC<ResultRenameDrawerProps> = (props) => {
-  const { className = '', nameMappingList, allDefNameMapping = {}, onClose, onAfterRenameTs, renameTypescript, ...otherProps } = props;
+  const {
+    className = '',
+    nameMappingList,
+    allDefNameMapping = {},
+    onClose,
+    onAfterRenameTs,
+    renameTypescript,
+    localServiceInfo = [],
+    ...otherProps
+  } = props;
 
   const { token } = theme.useToken();
   const [form] = Form.useForm<RenameMapping>();
@@ -94,12 +104,17 @@ const ResultRenameDrawer: React.FC<ResultRenameDrawerProps> = (props) => {
    * @param serviceName - 接口方法名称
    * @returns {Promise<void>} - 返回一个Promise对象，如果同一分组下的接口方法名称重复则抛出错误，否则解析为void
    */
-  const validateSameServiceName = (groupNamePath: NamePath, serviceName?: string) => {
+  const validateSameServiceName = (groupName: string, groupNamePath: NamePath, serviceName?: string) => {
     const currentGroup = (form.getFieldValue(groupNamePath) ?? []) as ApiGroupNameMapping[];
     const hasSameServiceName = currentGroup.filter((item) => item.serviceName === serviceName).length > 1;
 
     if (serviceName && hasSameServiceName) {
-      return Promise.reject(new Error('同一分组下的接口方法名称不能重复'));
+      return Promise.reject(new Error(`<${groupName}>分组下存在相同的接口方法名称，请重新命名`));
+    }
+
+    const localGroup = localServiceInfo.find((item) => item.groupName === groupName);
+    if (localGroup && localGroup.nameMappingList.some((item) => item.serviceName === serviceName)) {
+      return Promise.reject(new Error(`本地<${groupName}>分组下存在相同名称的接口文件，请重新命名`));
     }
 
     return Promise.resolve();
@@ -275,7 +290,8 @@ const ResultRenameDrawer: React.FC<ResultRenameDrawerProps> = (props) => {
                                         const isServiceName = nameField === 'serviceName'; // 是否接口名称
                                         if (isServiceName) {
                                           validationRule.push({
-                                            validator: (_, value) => validateSameServiceName([nameGroupFormKey, groupIndex, 'group'], value),
+                                            validator: (_, value) =>
+                                              validateSameServiceName(groupName, [nameGroupFormKey, groupIndex, 'group'], value),
                                           });
                                         }